+++ ochrona/eval/policy/models.py
@@ -1,0 +1,5 @@
+from collections import namedtuple
+
+TokenInstance = namedtuple("TokenInstance", ["id", "value"])
+
+Definition = namedtuple("Definition", ["field", "operator", "value"])